refactor(mobile): consolidate send success refresh events
