+++ apps/mobile/src/screens/Send/hooks/useSendToken.ts
@@ -59,11 +59,7 @@ import useAsyncFn from 'react-use/lib/useAsyncFn';
 import { useSwitchSceneAccountOnSelectedTokenWithOwner } from '@/databases/hooks/token';
 import { naviReplace } from '@/utils/navigation';
 import { RootNames } from '@/constant/layout';
-import {
-  useFocusEffect,
-  useIsFocused,
-  useRoute,
-} from '@react-navigation/native';
+import { useIsFocused, useRoute } from '@react-navigation/native';
 import { sendScreenParamsAtom } from '@/hooks/useSendRoutes';
 import { ITokenCheck } from '@/components/Token/TokenSelectorSheetModal';
 import {
@@ -71,7 +67,6 @@ import {
   makeAccountObject,
 } from '@/utils/account';
 import { usePollSendPendingCount } from './useSendPendingCount';
-import { eventBus, EventBusListeners, EVENTS } from '@/utils/events';
 import { useMemoizedFn } from 'ahooks';
 import {
   useRecentSendToHistoryFor,
@@ -1860,17 +1855,21 @@ export function useSendTokenForm({
     useRecentSendToHistoryFor(formValues.to);
 
   useEffect(() => {
-    const onTxCompleted: EventBusListeners[typeof EVENTS.TX_COMPLETED] =
-      txDetail => {
+    const disposeRets = [] as Function[];
+    subscribeEvent(
+      sendTokenEventsRef.current,
+      SendTokenEvents.ON_SIGNED_SUCCESS,
+      () => {
         reFetch();
         setTimeout(() => {
           reFetch();
         }, 5000);
-      };
-    eventBus.addListener(EVENTS.TX_COMPLETED, onTxCompleted);
+      },
+      { disposeRets },
+    );
 
     return () => {
-      eventBus.removeListener(EVENTS.TX_COMPLETED, onTxCompleted);
+      disposeRets.forEach(dispose => dispose());
     };
   }, [reFetch]);
 
@@ -1945,12 +1944,30 @@ export function useSendTokenForm({
     formik.resetForm();
   }, [setFormValues, formik]);
 
-  const refreshCurrentToken = useMemoizedFn(async () => {
-    return loadCurrentToken(
-      currentToken.id,
-      currentToken.chain,
-      currentAccount!.address,
-    );
+  const refreshCurrentTokenBalance = useMemoizedFn(async () => {
+    if (!currentAccount?.address) {
+      return;
+    }
+
+    putScreenState({
+      balanceError: null,
+      balanceWarn: null,
+    });
+    markBalanceLoading({
+      tokenId: currentToken.id,
+      chainId: currentToken.chain,
+      currentAddress: currentAccount.address,
+    });
+
+    try {
+      await loadCurrentToken(
+        currentToken.id,
+        currentToken.chain,
+        currentAccount.address,
+      );
+    } catch (error) {
+      console.error('SendScreen refresh current token error', error);
+    }
   });
 
   const prepareRef = useRef<Promise<Tx | void>>(undefined);
@@ -1999,6 +2016,23 @@ export function useSendTokenForm({
     handleFormValuesChange,
     formik.values,
   ]);
+
+  useEffect(() => {
+    const disposeRets = [] as Function[];
+    subscribeEvent(
+      sendTokenEventsRef.current,
+      SendTokenEvents.ON_SIGNED_SUCCESS,
+      () => {
+        resetAfterSignedSuccess();
+        refreshCurrentTokenBalance();
+      },
+      { disposeRets },
+    );
+
+    return () => {
+      disposeRets.forEach(dispose => dispose());
+    };
+  }, [refreshCurrentTokenBalance, resetAfterSignedSuccess]);
 
   const isFocused = useIsFocused();
 
@@ -2051,15 +2085,6 @@ export function useSendTokenForm({
     prepareDirectSubmitMiniTx,
   ]);
 
-  useFocusEffect(
-    useCallback(() => {
-      eventBus.addListener(EVENTS.RELOAD_TX, refreshCurrentToken);
-      return () => {
-        eventBus.removeListener(EVENTS.RELOAD_TX, refreshCurrentToken);
-      };
-    }, [refreshCurrentToken]),
-  );
-
   const svBottomAreaHeight = useSharedValue(220);
   useAnimatedReaction(
     () => {
@@ -2092,6 +2117,7 @@ export function useSendTokenForm({
 
     currentToken,
     loadCurrentToken,
+    refreshCurrentTokenBalance,
     checkCexSupport,
     handleCurrentTokenChange,
 
@@ -2114,7 +2140,6 @@ export function useSendTokenForm({
     formik,
     formValues,
     resetFormValues,
-    resetAfterSignedSuccess,
     handleFieldChange,
     patchFormValues,
     handleFormValuesChange,
